#20827 Fixed bug with listeners stopping on editing controller or exiting from rails console

+++ init.rb
@@ -9,46 +9,56 @@ Redmine::Plugin.register :rmplus_devtools do
   settings :partial => 'settings/rmplus_devtools'
 end
 
-Rails.application.config.after_initialize do
+def start_listeners
   if Rails.env.development?
-    enable_assets_listeners = Setting.plugin_rmplus_devtools[:enable_assets_listeners]
+    enable_assets_listeners = (Setting.plugin_rmplus_devtools || {})[:enable_assets_listeners] || false
     if enable_assets_listeners
-      $listeners = []
-      Rails.logger.debug "Initializing listeners..."
-      Redmine::Plugin.registered_plugins.each do |name, plugin|
-        source = plugin.assets_directory
-        if File.exist?(source) && File.directory?(source)
-          destination = plugin.public_directory
-          assets_listener = Listen.to source do |modified, added, removed|
-            modified.each do |modified_path|
-              if File.file?(modified_path)
-                target = File.join(destination, modified_path.gsub(source, ''))
-                FileUtils.cp(modified_path, target)
+      if $listeners.nil? || ($listeners.is_a?(Array) && $listeners.empty?)
+        $listeners = []
+        Rails.logger.debug "Initializing listeners..."
+        Redmine::Plugin.registered_plugins.each do |name, plugin|
+          source = plugin.assets_directory
+          if File.exist?(source) && File.directory?(source)
+            destination = plugin.public_directory
+            assets_listener = Listen.to source do |modified, added, removed|
+              modified.each do |modified_path|
+                if File.file?(modified_path)
+                  target = File.join(destination, modified_path.gsub(source, ''))
+                  FileUtils.cp(modified_path, target)
+                end
+              end
+              added.each do |added_path|
+                if File.directory?(added_path)
+                  FileUtils.mkdir_p(added_path)
+                elsif File.file?(added_path)
+                  target = File.join(destination, added_path.gsub(source, ''))
+                  FileUtils.cp(added_path, target)
+                end
+              end
+              removed.each do |removed_path|
+                target = File.join(destination, removed_path.gsub(source, ''))
+                FileUtils.remove_entry(target, true)
               end
             end
-            added.each do |added_path|
-              if File.directory?(added_path)
-                FileUtils.mkdir_p(added_path)
-              elsif File.file?(added_path)
-                target = File.join(destination, added_path.gsub(source, ''))
-                FileUtils.cp(added_path, target)
-              end
-            end
-            removed.each do |removed_path|
-              target = File.join(destination, removed_path.gsub(source, ''))
-              FileUtils.remove_entry(target, true)
-            end
+            assets_listener.start
+            $listeners << assets_listener
           end
-          Rails.logger.debug "Starting assets listener for plugin #{name} in directory #{source}"
-          assets_listener.start
-          $listeners << assets_listener
         end
       end
       at_exit do
-        Rails.logger.debug "Stopping listeners..."
-        $listeners.each{ |listener| listener.stop }
+        if not defined?(Rails::Console)
+          Rails.logger.debug "Stopping listeners..."
+          $listeners.each{ |listener| listener.stop }
+        end
       end
     end
   end
+end
 
+Rails.application.config.after_initialize do
+  start_listeners
+end
+
+Rails.application.config.to_prepare do
+  start_listeners
 end
